--- a/vapid/soa.h
+++ b/vapid/soa.h
@@ -32,7 +32,7 @@ namespace vapid {
         TupleDumper<std::integral_constant<uint16_t, 0>, T>::dump(ss, t);
     }
 
-    template <typename... Ts>
+    template <bool prefer_speed, typename... Ts>
     class soa {
     public:
         using backing_type = std::tuple<std::vector<Ts>...>;
@@ -50,7 +50,7 @@ namespace vapid {
 
         template<size_t col_idx>
         nth_col_type<col_idx>& get_column() {
-            const soa<Ts...>* const_this = this;
+            const soa<prefer_speed, Ts...>* const_this = this;
             return const_cast<nth_col_type<col_idx>&>(const_this->get_column<col_idx>());
         }
 
@@ -65,6 +65,8 @@ namespace vapid {
         template <typename... Xs>
         void insert(Xs... xs) {
             insert_impl(std::index_sequence_for<Ts...>{}, std::forward_as_tuple(xs...));
+            if constexpr (prefer_speed)
+               sort_order_reference_.push_back(size() - 1);
         }
 
         auto get_row(size_t row) const {
@@ -107,30 +109,42 @@ namespace vapid {
 
         template <size_t col_idx, typename C>
         void sort_by_field(C&& comparator) {
-            size_t num_elems =  size();
-            thread_local std::vector<size_t> sort_order_tmp;
-            thread_local std::vector<size_t> sort_order_reference;
+            if constexpr (prefer_speed) {
+               reset_sort_reference();
+               auto& col = get_column<col_idx>();
 
-            sort_order_tmp.resize(num_elems);
-            sort_order_reference.resize(num_elems);
-            for (size_t i = 0; i < num_elems; ++i) {
-                sort_order_tmp[i] = i;
+               auto comparator_wrapper = [=](size_t a, size_t b) {
+                   return comparator(col[a], col[b]);
+               };
+
+                std::stable_sort(sort_order_reference_.begin(),
+                     sort_order_reference_.end(),
+                     comparator_wrapper);
+               sort_by_reference_impl(std::index_sequence_for<Ts...>{});
+             } else {
+                size_t num_elems =  size();
+                thread_local std::vector<size_t> sort_order_tmp;
+                thread_local std::vector<size_t> sort_order_reference;
+                
+                sort_order_tmp.resize(num_elems);
+                sort_order_reference.resize(num_elems);
+                for (size_t i = 0; i < num_elems; ++i) {
+                    sort_order_tmp[i] = i;
+                }
+                auto& col = get_column<col_idx>();
+
+                auto comparator_wrapper = [=](size_t a, size_t b) {
+                    return comparator(col[a], col[b]);
+                };
+
+                std::stable_sort(sort_order_tmp.begin(),
+                                 sort_order_tmp.end(),
+                                 comparator_wrapper);
+                for (size_t i=0; i<num_elems; ++i)
+                   sort_order_reference[sort_order_tmp[i]] = i;
+            
+                sort_by_reference_impl(sort_order_tmp, sort_order_reference, std::index_sequence_for<Ts...>{});
             }
-
-            auto& col = get_column<col_idx>();
-
-            auto comparator_wrapper = [=](size_t a, size_t b) {
-                return comparator(col[a], col[b]);
-            };
-
-            std::stable_sort(sort_order_tmp.begin(),
-                             sort_order_tmp.end(),
-                             comparator_wrapper);
-
-            for (size_t i=0; i<num_elems; ++i)
-                sort_order_reference[sort_order_tmp[i]] = i;
-
-            sort_by_reference_impl(sort_order_tmp, sort_order_reference, std::index_sequence_for<Ts...>{});
         }
 
         template <size_t col_idx>
@@ -157,7 +171,14 @@ namespace vapid {
             ss << "}" << std::endl;
         }
 
-        void prepare_tmp() {}
+        void prepare_tmp() {
+            if constexpr (prefer_speed) {
+                // sorting requires the temporary buffers in data_tmp_
+                // this function pre-allocates those temporary buffers
+                // so that no allocation is done during the sort call
+                resize_impl(std::index_sequence_for<Ts...>{}, data_tmp_, size());  
+            }
+        }
 
     private:
         template <typename T, size_t... I>
@@ -190,11 +211,24 @@ namespace vapid {
             ((get_column<I>().reserve(res_size)), ...);
         }
 
+        void reset_sort_reference() {
+            if constexpr (prefer_speed) {
+                for (size_t i = 0; i < size(); ++i) {
+                    sort_order_reference_[i] = i;
+                }
+            }
+        }
+
         template <size_t... I>
         void sort_by_reference_impl(std::vector<size_t> &sort_order_tmp,
                                     const std::vector<size_t> &sort_order_reference,
                                     std::integer_sequence<size_t, I...>) {
             ((sort_col_by_reference(sort_order_tmp, sort_order_reference, std::integral_constant<size_t, I>{})), ...);
+        }
+
+        template <size_t... I>
+        void sort_by_reference_impl(std::integer_sequence<size_t, I...>) {
+            ((sort_col_by_reference(std::integral_constant<size_t, I>{})), ...);
         }
 
         
@@ -225,11 +259,33 @@ namespace vapid {
             reorder(col, sort_order_tmp);
         }
 
+        template <size_t col_idx>
+        void sort_col_by_reference(std::integral_constant<size_t, col_idx>) {
+            auto& src = std::get<col_idx>(data_);
+            auto& dst = std::get<col_idx>(data_tmp_);
+            dst.resize(src.size());
+            for (size_t idx = 0; idx < src.size(); ++idx) {
+                dst[idx] = std::move(src[sort_order_reference_[idx]]);
+            }
+            std::swap(src, dst);
+        }
+
         std::tuple<std::vector<Ts>...> data_;
+
+        struct empty_ { };
+        using ts = std::conditional_t<prefer_speed, std::tuple<std::vector<Ts>...>, empty_>;
+        using cond_size_vec = std::conditional_t<prefer_speed, std::vector<size_t>, empty_>;
+
+        
+        // tmp buffers for reordering when sorting
+        ts data_tmp_;
+
+        // the reference permutation describing sorted order
+        cond_size_vec sort_order_reference_;
     };
 
-    template <typename... Ts>
-    std::ostream& operator<<(std::ostream& cout, const vapid::soa<Ts...>& soa) {
+    template <bool prefer_speed, typename... Ts>
+    std::ostream& operator<<(std::ostream& cout, const vapid::soa<prefer_speed, Ts...>& soa) {
         soa.dump(cout);
         return cout;
     }