@markisus 2022-02-09
Thanks so much for the contribution! I did some cursory testing against the benchmark.cc and it seems it might be slower than the code on master, at least in visual studio 2022. I'll integrate the google benchmark library this weekend so we can get some more solid numbers.