@markisus 2022-02-10
Thanks for all the work you've put into this. I've recently integrated a more robust benchmark (benchmarks.cc with an s) and this PR does not compare favorably to master. See the "SoaSortBySensorId" benchmark. This is with MSCV. I would try gcc as well, but there is some sort of pessimization going on that blows up the sorting time that I haven't figured out yet.

```
# This PR

Run on (8 X 2995 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 1280 KiB (x4)
  L3 Unified 12288 KiB (x1)
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_SoaSortBySensorId   15075959 ns     14100610 ns           41
BM_VecSortBySensorId   23261811 ns     22321429 ns           28
BM_SoaSumTimestamps      111306 ns       111607 ns         5600
BM_VecSumTimestamps      286398 ns       285993 ns         2240
```

```
# master

Run on (8 X 2995 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 1280 KiB (x4)
  L3 Unified 12288 KiB (x1)
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_SoaSortBySensorId   10599404 ns      9151786 ns           70
BM_VecSortBySensorId   24412036 ns     25111607 ns           28
BM_SoaSumTimestamps      106078 ns       106027 ns         5600
BM_VecSumTimestamps      264606 ns       266841 ns         2635
```