@markisus 2022-02-10
Nice catch. I updated master to use std::move and re-ran the new benchmarks.cc with string and array sensor data. This change does not affect the array sensor data, since moving cannot be done without copying.
```
# master
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_SoaSortBySensorId_ArrayData    10218730 ns     10625000 ns           50
BM_VecSortBySensorId_ArrayData    23377407 ns     23437500 ns           30
BM_SoaSortBySensorId_StringData   12121888 ns     11160714 ns           56
BM_VecSortBySensorId_StringData   15877022 ns     15243902 ns           41
BM_SoaSumTimestamps_ArrayData       102751 ns       102534 ns         7467
BM_VecSumTimestamps_ArrayData       267027 ns       269938 ns         2489
```

Here are the times with the updated benchmark after merging your branch. (resolving conflict by taking your code)
```
Run on (8 X 2995 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 1280 KiB (x4)
  L3 Unified 12288 KiB (x1)
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_SoaSortBySensorId_ArrayData    15067329 ns     15066964 ns           56
BM_VecSortBySensorId_ArrayData    24415282 ns     24553571 ns           28
BM_SoaSortBySensorId_StringData   17082551 ns     16387195 ns           41
BM_VecSortBySensorId_StringData   16654598 ns     16768293 ns           41
BM_SoaSumTimestamps_ArrayData       103921 ns       102534 ns         7467
BM_VecSumTimestamps_ArrayData       254620 ns       257382 ns         2489
```

For both array and string data, SoaSort is faster on master.