@greg7mdp 2022-02-10
strange that my branch shows more iterations on the first line?

> For both array and string data, SoaSort is faster on master.

Yes but it uses double the memory of the vector. This could be an issue for very large datasets if we run out of memory.