@markisus 2022-02-10
I think there is some runtime calculation by the benchmarking library that
smartly determines how many iterations to run to get a good estimate. I
would pay attention only to the time column.

Could you put your feature as a constructor option? Then we could still
have the d behavior as well and I'd be happy to merge it.

On Thu, Feb 10, 2022, 12:49 PM Gregory Popovitch ***@***.***>
wrote:

> strange that my branch shows more iterations on the first line?
>
> For both array and string data, SoaSort is faster on master.
>
> Yes but it uses double the memory of the vector. This could be an issue
> for very large datasets if we run out of memory.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/markisus/vapid-soa/pull/3#issuecomment-1035499317>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADSVOJ5ZRRYYN5M5BAYZATU2QQENANCNFSM5N4FULKQ>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
> You are receiving this because you commented.Message ID:
> ***@***.***>
>