@greg7mdp 2022-02-11
I did it with a template parameter, so that there is no overhead. What do you think?