@markisus 2022-02-11
For my eyes it's kind of cluttering to the type signature.