@greg7mdp 2022-02-11
> it's only a size_t and a pointer per field if nothing is allocated

Hum, no, a vector is 3 pointers, so it is 3 pointers per firel + 3 pointers that we are adding even if nothing is allocated.

> Is there a way to make a default if the user does not specify? 

Yes, that was my first thought, but I don't think it is possible.

I understand if you don't want to merge it, I'm fine if you close the PR, it was interesting to look into this.