@markisus 2022-02-11
Alright thanks for the discussion. Very enlightening. I did not know that a vector was 3 pointers. I thought you would only have to store the size and a pointer to the head of the underlying memory.