@markisus 2022-02-11
@greg7mdp You may be interested in this implementation https://github.com/markisus/vapid-soa/pull/4 in which I remove the need for your `sort_order_tmp` by doing some pre-analysis on the permutation. I also eat the overhead of unused std::vects in order to preserve type signature.