@greg7mdp 2022-02-11
Thanks for letting me know, @markisus. I have also worked on a simpler version on my side which I think is similar, but more straightforward than yours. It is faster than my previous version, but still not as fast as your duplicated vectors.
You can see it at: https://github.com/greg7mdp/vapid-soa/blob/master/vapid/soa.h