@markisus 2022-02-11
I see, very nice! I think the only difference is I memorize the number of swaps on each inner loop and you cheaply compute it on the fly using your bit vector.